+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 ## Unreleased
+
 * Added
   * Added examples for the Xaxis Overlapping tick values.